refactor radio page composition
